Rewrite IconButton
