+++ packages/bento-design-system/src/index.ts
@@ -25,7 +25,7 @@ export * from "./Divider/Divider";
 export * from "./Feedback/createFeedback";
 export * from "./Field/createFormFields";
 export * from "./Form/createFormLayoutComponents";
-export * from "./IconButton/createIconButton";
+export * from "./IconButton/IconButton";
 export * from "./Icons";
 export * from "./Icons/svgIconProps";
 export * from "./Illustrations";
@@ -78,7 +78,7 @@ export { inputRecipe } from "./Field/Field.css";
 
 export type { SelectFieldProps } from "./SelectField/createSelectField";
 export type { FieldProps } from "./Field/FieldProps";
-export type { IconButtonProps } from "./IconButton/createIconButton";
+export type { IconButtonProps } from "./IconButton/IconButton";
 
 export * from "./util/atoms";
 export * from "./util/breakpoints";
